+++ test/container/container.test.ts
@@ -1125,4 +1125,31 @@ describe("Container", () => {
         expect(weapon).to.be.instanceOf(Shuriken);
     });
 
+    it("container resolve should come from the same container", () => {
+        @injectable()
+        class CompositionRoot{}
+        class DerivedContainer extends Container{
+            public planningForCompositionRoot(): void {
+                //
+            }
+        }
+        const middleware:interfaces.Middleware = (next) =>
+            (nextArgs) => {
+                const contextInterceptor = nextArgs.contextInterceptor;
+                nextArgs.contextInterceptor = context => {
+                    if(context.plan.rootRequest.serviceIdentifier === CompositionRoot){
+                        (context.container as DerivedContainer).planningForCompositionRoot();
+                    }
+                    return contextInterceptor(context);
+                }
+                return next(nextArgs)
+            }
+
+        const myContainer = new DerivedContainer();
+        myContainer.applyMiddleware(middleware);
+        myContainer.resolve(CompositionRoot);
+        // tslint:disable-next-line: no-unused-expression
+        expect(() => myContainer.resolve(CompositionRoot)).not.to.throw;
+    })
+
 });